refactor(plugin): gatsby-image module
