+++ packages/gatsby-transformer-cloudinary/api/index.js
@@ -1,5 +1,10 @@
+const {
+  getFixedImageObject,
+  getFluidImageObject,
+} = require('../gatsby-image/get-image-objects');
+
 // Runtime APIs for fetching remote images on Cloudinary
-exports.getFixedImageObject =
-  require('../get-image-objects').getFixedImageObject;
-exports.getFluidImageObject =
-  require('../get-image-objects').getFluidImageObject;
+module.exports = {
+  getFixedImageObject,
+  getFluidImageObject,
+};
