+++ packages/gatsby-transformer-cloudinary/gatsby-image/get-shared-image-data.test.js
