+++ packages/gatsby-transformer-cloudinary/gatsby-node.js
@@ -1,8 +1,4 @@
 const fs = require('fs-extra');
-const {
-  getFixedImageObject,
-  getFluidImageObject,
-} = require('./get-image-objects');
 const { uploadImageNodeToCloudinary } = require('./upload');
 const { setPluginOptions } = require('./options');
 const { createImageNode } = require('./create-image-node');
@@ -13,6 +9,8 @@ const {
 //import plugin options, used to check for API key before uploading assets to Cloudinary
 const { getPluginOptions } = require('./options');
 const pluginOptions = getPluginOptions();
+
+const { createGatsbyImageResolvers } = require('./gatsby-image');
 
 const ALLOWED_MEDIA_TYPES = ['image/png', 'image/jpeg', 'image/gif'];
 
@@ -78,107 +76,9 @@ exports.createSchemaCustomization = ({ actions }) => {
   `);
 };
 
-exports.createResolvers = ({ createResolvers, reporter }) => {
-  const resolvers = {
-    CloudinaryAsset: {
-      fixed: {
-        type: 'CloudinaryAssetFixed!',
-        resolve: (
-          {
-            public_id,
-            version,
-            cloudName,
-            originalHeight,
-            originalWidth,
-            defaultBase64,
-            defaultTracedSVG,
-          },
-          {
-            base64Width,
-            base64Transformations,
-            ignoreDefaultBase64,
-            height,
-            width,
-            transformations,
-            chained,
-          },
-          _context,
-          info
-        ) => {
-          const fieldsToSelect = info.fieldNodes[0].selectionSet.selections.map(
-            (item) => item.name.value
-          );
-          return getFixedImageObject({
-            base64Transformations,
-            base64Width,
-            chained,
-            cloudName,
-            defaultBase64,
-            fieldsToSelect,
-            defaultTracedSVG,
-            height,
-            ignoreDefaultBase64,
-            originalHeight,
-            originalWidth,
-            public_id,
-            reporter,
-            transformations,
-            version,
-            width,
-          });
-        },
-      },
-      fluid: {
-        type: 'CloudinaryAssetFluid!',
-        resolve: (
-          {
-            breakpoints,
-            cloudName,
-            defaultBase64,
-            defaultTracedSVG,
-            originalHeight,
-            originalWidth,
-            public_id,
-            version,
-          },
-          {
-            base64Transformations,
-            base64Width,
-            chained,
-            ignoreDefaultBase64,
-            maxWidth,
-            transformations,
-          },
-          _context,
-          info
-        ) => {
-          const fieldsToSelect = info.fieldNodes[0].selectionSet.selections.map(
-            (item) => item.name.value
-          );
-          return getFluidImageObject({
-            base64Transformations,
-            base64Width,
-            breakpoints,
-            chained,
-            cloudName,
-            defaultBase64,
-            fieldsToSelect,
-            defaultTracedSVG,
-            ignoreDefaultBase64,
-            maxWidth,
-            originalHeight,
-            originalWidth,
-            public_id,
-            reporter,
-            transformations,
-            version,
-          });
-        },
-      },
-    },
-  };
-
-  createResolvers(resolvers);
+exports.createResolvers = (gatsbyUtils) => {
+  // To be used with gatsby-image
+  createGatsbyImageResolvers(gatsbyUtils);
 };
 
 async function createAssetNodeFromFile({
