--- packages/gatsby-transformer-cloudinary/get-image-objects/get-aspect-ratio.js
@@ -1,16 +1,0 @@
-// retrieve aspect ratio if in transformation else create aspect ratio values
-exports.getAspectRatio = (transformations, originalAspectRatio) => {
-  const arTransform = transformations.find((t) => t.startsWith('ar_'));
-  if (!arTransform) {
-    return originalAspectRatio;
-  }
-
-  const newAspectRatio = arTransform.replace('ar_', '');
-  if (newAspectRatio.indexOf(':') === -1) {
-    return Number(newAspectRatio);
-  }
-
-  const [w, h] = newAspectRatio.split(':').map(Number);
-
-  return w / h;
-};
